Work on software controlled fast exposures
Also a bunch of performance enhancements to the software exposure path.
Before these changes, the fastest software exposure was about 800ms. It
will now go to just a few ms (although there is about 1.5 ms of noise,
so really short exposures are inaccurate).

Because of this, I added code to use the hardware timer for exposures
less than 1 second for the main camera. The guide camera already
uses the hardware timer for short exposures.