Fixing a try/finally error
That was introduced when all the functios were wrapped with
try/catch blocks.  A try/catch was changed to a try/finally
block, but that was not correct.

The bad change caused state to be unconditionally set to idle at
the end of StartExposure(), insteat of just when an exception
occured.