Seperate the config date for the cameras
All the configuratino data used to be attached to SXMain because
at one time that camera was guaranteed to be present, but when
the setup wizard was added, it became possible to haave
configurations which did not have one. This required rewriting
the configuration code so that the configuration values are
attached to each individual camera.