Set default max bining to 4
This works around a problem I was having with conform.  The SX
cameras support binning 1X-8X, execpt for 7X.  There is no way
to specify this behavior with the current ASCOM spec.  If I
make the driver reject 7X, conform complains.  So I set the
default max to 4, conform is happy, and if a user really wants
8X, they can enable it.  I don't like it, but it was the cleanest
solution I could find.