Adding a MaxIoSize
There have been some issues with the driver working on some systems
and not on others. I'm adding a way to limit the I/O sizes, in
case it is I/O size related.

The current value is very large (64M), but with it in the code,
it can be turned down for debugging as needed.